Introduce a shared project for the showcase and user admin websites
Was pulled from the UserAdmin Domain project enumeration into the Shared.
