+++ src/Shared/Smart.FA.Catalog.Shared/Smart.FA.Catalog.Shared.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\UserAdmin\src\Smart.FA.Catalog.Localization\Smart.FA.Catalog.Localization.csproj" />
+  </ItemGroup>
+
+</Project>
