+++ src/UserAdmin/src/Smart.FA.Catalog.Application/UseCases/Commands/CreateTrainingCommandHandler.cs
@@ -7,6 +7,7 @@ using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Core.Domain.Interfaces;
 using Smart.FA.Catalog.Core.LogEvents;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Training;
 
 namespace Smart.FA.Catalog.Application.UseCases.Commands;
