+++ src/UserAdmin/src/Smart.FA.Catalog.Core/Domain/Trainer/PersonalSocialNetwork.cs
@@ -1,5 +1,6 @@
 using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Trainer;
 
 namespace Smart.FA.Catalog.Core.Domain;
