+++ src/UserAdmin/src/Smart.FA.Catalog.Core/Domain/Trainer/Trainer.cs
@@ -1,6 +1,7 @@
 using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Core.Domain.ValueObjects;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Trainer;
 
 namespace Smart.FA.Catalog.Core.Domain;
