+++ src/UserAdmin/src/Smart.FA.Catalog.Core/Domain/Training/Specifications/TrainingSpecifications.cs
@@ -1,6 +1,7 @@
 using System.Linq.Expressions;
 using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Training;
 
 namespace Smart.FA.Catalog.Core.Domain.Specifications;
