+++ src/UserAdmin/src/Smart.FA.Catalog.Core/Domain/Training/TrainingTarget.cs
@@ -1,4 +1,5 @@
 using Smart.FA.Catalog.Core.Domain.Enumerations;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Training;
 
 namespace Smart.FA.Catalog.Core.Domain;
