+++ src/UserAdmin/src/Smart.FA.Catalog.Infrastructure/Persistence/Configuration/TrainerPersonalNetworkConfiguration.cs
@@ -4,6 +4,8 @@ using Smart.FA.Catalog.Core.Domain;
 using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Core.SeedWork;
 using Smart.FA.Catalog.Infrastructure.Persistence.Configuration.EntityConfigurations;
+using Smart.FA.Catalog.Shared.Domain.Enumerations;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Trainer;
 
 namespace Smart.FA.Catalog.Infrastructure.Persistence.Configuration;
