+++ src/UserAdmin/src/Smart.FA.Catalog.Infrastructure/Persistence/Configuration/TrainingSlotNumberTypeConfigurations.cs
@@ -1,5 +1,6 @@
 using Smart.FA.Catalog.Core.Domain.Enumerations;
 using Smart.FA.Catalog.Infrastructure.Persistence.Configuration.EntityConfigurations;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Training;
 
 namespace Smart.FA.Catalog.Infrastructure.Persistence.Configuration;
