+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Extensions/Middlewares/ProxyHeaderMiddleware.cs
@@ -5,6 +5,7 @@ using Smart.FA.Catalog.Application.UseCases.Queries;
 using Smart.FA.Catalog.Core.Domain.Models;
 using Smart.FA.Catalog.Core.Domain.User.Enumerations;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations;
 
 namespace Smart.FA.Catalog.Web.Extensions.Middlewares;
