+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Admin/Trainings/Create/Index.cshtml
@@ -2,6 +2,8 @@
 @using Microsoft.Extensions.Localization
 @using Smart.FA.Catalog.Core.Domain.Enumerations
 @using Smart.FA.Catalog.Core.SeedWork
+@using Smart.FA.Catalog.Shared.Domain.Enumerations
+@using Smart.FA.Catalog.Shared.Domain.Enumerations.Training
 @inject IStringLocalizer<CatalogResources> _localizer
 @model Smart.FA.Catalog.Web.Pages.Admin.Trainings.Create.CreateModel
 @{
