+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Admin/Trainings/List/TrainingListingViewModel.cs
@@ -1,5 +1,6 @@
 using Smart.Design.Razor.TagHelpers.Pill;
 using Smart.FA.Catalog.Core.Domain.Enumerations;
+using Smart.FA.Catalog.Shared.Domain.Enumerations.Training;
 
 namespace Smart.FA.Catalog.Web.Pages.Admin.Trainings.List;
