+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Admin/Trainings/Update/Index.cshtml
@@ -2,6 +2,8 @@
 @using Microsoft.Extensions.Localization
 @using Smart.FA.Catalog.Core.Domain.Enumerations
 @using Smart.FA.Catalog.Core.SeedWork
+@using Smart.FA.Catalog.Shared.Domain.Enumerations
+@using Smart.FA.Catalog.Shared.Domain.Enumerations.Training
 @using Smart.FA.Catalog.Web.Extensions
 
 @inject IStringLocalizer<CatalogResources> _localizer
