+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Pages/Shared/Components/AdminTrainingTile/Default.cshtml
@@ -1,6 +1,8 @@
 ﻿@using Microsoft.Extensions.Localization
 @using Smart.FA.Catalog.Core.Domain.Enumerations
 @using Smart.FA.Catalog.Core.SeedWork
+@using Smart.FA.Catalog.Shared.Domain.Enumerations
+@using Smart.FA.Catalog.Shared.Domain.Enumerations.Training
 
 @inject IStringLocalizer<CatalogResources> _localizer
 @model IGrouping<int, Core.Domain.Dto.TrainingDto>
