+++ src/UserAdmin/test/Smart.FA.Catalog.Tests.Common/TrainerFactory.cs
@@ -5,6 +5,7 @@ using Smart.FA.Catalog.Core.Domain.User.Dto;
 using Smart.FA.Catalog.Core.Domain.User.Enumerations;
 using Smart.FA.Catalog.Core.Domain.ValueObjects;
 using Smart.FA.Catalog.Core.SeedWork;
+using Smart.FA.Catalog.Shared.Domain.Enumerations;
 
 namespace Smart.FA.Catalog.Tests.Common;
